+++ python/eyes_universal/applitools/eyes_universal/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = "2.16.0"
+__version__ = "2.15.0"
 
 
 def get_instance():